zippo_single/requirements.txt
APScheduler==2.1.2
certifi==2018.4.16
chardet==3.0.4
diff-match-patch==20121119
Django==1.9
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-simple-captcha==0.4.6
et-xmlfile==1.0.1
future==0.16.0
geoip2==2.8.0
httplib2==0.11.3
idna==2.6
jdcal==1.4
maxminddb==1.3.0
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.3
paho-mqtt==1.3.1
Pillow==5.1.0
PyMySQL==0.8.1
pytz==2018.4
PyYAML==3.12
requests==2.18.4
schedule==0.5.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
